Fixed Errors And Performed Optimizations
I fixed code that led to errors by changing it a little. In addition, I got rid of empty println statements and added newline escape characters to the rest of them. As a result of that, the test execution time was reduced by 0.03 seconds.

Originally the execution time was 0.096 seconds, and I got it down to 0.066 seconds.